requirements.txt
Django>=3.2.4,<3.3 #install latest patch version before 3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16 # used for testing
django-cors-headers>=3.13.0,<3.14
# djangorestframework-queryfields
